Refactored AbstractJpaQuery to use the same TupleConverter used in SimpleJpaRepository
